[DatabaseUserType] move to model package too verbose already, methinks
