+++ database_mockup/src/animalkeeping/mappings/DatabaseUser.hbm.xml
@@ -15,7 +15,7 @@
         </id>
         <property name="name" type="string" column="name" not-null="true" unique="true"/>
         <one-to-one name="person"/>
-        <many-to-one name="type" column="type_id" class="animalkeeping.technicalAdministration.DatabaseUserType" not-null="true"/>
+        <many-to-one name="type" column="type_id" class="animalkeeping.model.DatabaseUserType" not-null="true"/>
     </class>
 
 </hibernate-mapping>
