+++ database_mockup/src/animalkeeping/mappings/DatabaseUserType.hbm.xml
@@ -4,7 +4,7 @@
         "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
         "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
 
-<hibernate-mapping package="animalkeeping.technicalAdministration">
+<hibernate-mapping package="animalkeeping.model">
     <class name="DatabaseUserType" table="DatabaseUserType">
         <id name="id" column="id">
             <generator class="increment"/>
