+++ database_mockup/src/animalkeeping/ui/SuperUserForm.java
@@ -1,8 +1,5 @@
 package animalkeeping.ui;
 
-import animalkeeping.technicalAdministration.DatabaseUserType;
-import animalkeeping.util.AddDatabaseUserDialog;
-import javafx.scene.control.ComboBox;
 import javafx.scene.control.Label;
 import javafx.scene.control.PasswordField;
 import javafx.scene.control.TextField;
@@ -10,15 +7,10 @@ import javafx.scene.layout.ColumnConstraints;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
-import javafx.util.StringConverter;
-import org.hibernate.HibernateException;
-import org.hibernate.Session;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
 
 import static animalkeeping.util.Dialogs.showInfo;
